♻️ chore(prompter): move questions to internal module
