+++ src/question_prompter/dependencies_step.py
@@ -1,6 +1,6 @@
-from src.question_prompter.boolean_question import BooleanQuestion
-from src.question_prompter.choice_question import ChoiceQuestion
-from src.question_prompter.multiple_choice_question import MultipleChoiceQuestion
+from src.question_prompter.question.boolean_question import BooleanQuestion
+from src.question_prompter.question.choice_question import ChoiceQuestion
+from src.question_prompter.question.multiple_choice_question import MultipleChoiceQuestion
 from src.question_prompter.steps import Step
