+++ src/question_prompter/general_project_step.py
@@ -1,6 +1,6 @@
-from src.question_prompter.boolean_question import BooleanQuestion
-from src.question_prompter.choice_question import ChoiceQuestion
-from src.question_prompter.free_text_question import FreeTextQuestion
+from src.question_prompter.question.boolean_question import BooleanQuestion
+from src.question_prompter.question.choice_question import ChoiceQuestion
+from src.question_prompter.question.free_text_question import FreeTextQuestion
 from src.question_prompter.steps import Step
